feat: implement test environment management - complete task 4.2
- TemporaryDirectory class provides isolated test environments with automatic cleanup
- TemporaryCache class creates temporary cache environments for testing
- TestFixture base class manages setup/teardown lifecycle with helper methods
- TestDataManager creates realistic solar system test database with 27+ celestial bodies
- Comprehensive test data loading for JPL responses, ephemeris data, and cache samples
- Automatic cleanup mechanisms prevent test data leakage
- All functionality verified with comprehensive testing

Requirements satisfied: 7.4, 7.5

+++ lib/solar_core/include/solar_core/bodies/celestial_body.hpp
@@ -87,8 +87,8 @@ class CelestialBody {
 
   // Gravitational interactions
   [[nodiscard]] Math::Vector3d gravitational_force_to(const CelestialBody& other) const noexcept;
-  [[nodiscard]] double distance_to(const CelestialBody& other) const noexcept;
-  [[nodiscard]] double distance_squared_to(const CelestialBody& other) const noexcept;
+  [[nodiscard]] long double distance_to(const CelestialBody& other) const noexcept;
+  [[nodiscard]] long double distance_squared_to(const CelestialBody& other) const noexcept;
 
   // Utility functions
   [[nodiscard]] bool is_available_at(std::chrono::system_clock::time_point time) const noexcept;
